+++ app/src/main/ui/src/components/users/UserForm.tsx
@@ -82,6 +82,7 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
     return <>
         <ConfirmationDialog
             open={open ?? false}
+            data-testid="remove-group-dialog"
             onClose={() => toggleModal(false)}
             onConfirm={async () => {
                 await removeGroup(groupName ?? "");
@@ -127,7 +128,7 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
             }
 
             <div className={"my-4"}>
-                <Button variant={"primary"} type={"submit"} disabled={!handler.dirty}>
+                <Button variant={"primary"} type={"submit"} disabled={!handler.dirty} data-testid="save-button">
                     {t("editUser.buttons.save")}
                 </Button>
                 {props.additionalButtons as JSX.Element}