+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rewrite frontend with React (#77)
 - Extend search endpoint to pass a list of excluded values (#77)
 - Update documentation with current screenshots and better explanations (#81)
+- Update cypress-version for integration-tests (#78)
+
+### Added
+- New integration-tests for rewritten frontend (#78)
 
 ## [v1.9.0-1] - 2022-11-14
 ### Changed